requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
jupyter
notebook
kaggle
transformers
torch
pyspellchecker
streamlit
pyspellchecker
flask
